add Number Increase Animation

+++ SKS_Collection/MainController/LayerAnimation/View/CicleProgressAnimationView.h
@@ -10,6 +10,6 @@
 
 @interface CicleProgressAnimationView : BaseLayerAnimationView
 
-- (void)showWithProgress:(CGFloat)progress;
+- (void)showWithProgress:(int64_t)progress;
 
 @end
